Actualizamos para que sea compatible con usuarios

+++ router/publications.py
@@ -1,12 +1,14 @@
-from fastapi import APIRouter, HTTPException, status, Form
+from fastapi import APIRouter, HTTPException, status, Form, Depends
 from datetime import datetime, timezone
 from typing import List, Optional
 from bson import ObjectId
 
 from db.client import db
 from db.models.publication import Publication
+from db.models.user import User
 from db.schemas.publication import publicationSchema
 from utils.publicationsUtils import srcPublicationId
+from utils.authUtils import verifyTokenAccess
 
 router = APIRouter(
     prefix="/pb",
@@ -14,9 +16,10 @@ router = APIRouter(
 )
 
 @router.post("/create", status_code=status.HTTP_201_CREATED)
-async def createPb(title:str = Form(...), content:str = Form(...), category:str = Form(...), tags:List[str] = Form(...)):
+async def createPb(title:str = Form(...), content:str = Form(...), category:str = Form(...), tags:List[str] = Form(...), auth:User = Depends(verifyTokenAccess)):
     try:
         pb = {
+            "userId": auth.id,
             "title":title,
             "content":content,
             "category":category,
@@ -31,10 +34,13 @@ async def createPb(title:str = Form(...), content:str = Form(...), category:str 
         raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail={"message":"error: cannot create de publication"})
 
 @router.get("/get/{id}", status_code=status.HTTP_200_OK)
-async def getPb(id:str):
+async def getPb(id:str, auth:User = Depends(verifyTokenAccess)):
     try:
         publicationDb = srcPublicationId(ObjectId(id))
-        return Publication(**publicationDb)
+        if publicationDb["userId"] == auth.id:
+            return Publication(**publicationDb)
+        else:
+            raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to access this publication")
     except HTTPException as exception:
         if exception.status_code == status.HTTP_404_NOT_FOUND:
             raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Publication not found")
@@ -42,11 +48,11 @@ async def getPb(id:str):
             raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail={"message":"error: cannot get de publication"})
 
 @router.get("/all", status_code=status.HTTP_200_OK)
-async def getAllPb():
+async def getAllPb(auth:User = Depends(verifyTokenAccess)):
     try:
-        pbList = List[Publication]
-        pbList = []
-        publications = db.publications.find()
+        if auth.is_superuser:
+            pbList:List[Publication] = []
+            publications = db.publications.find()
         for publication in publications:
             publication["_id"] = str(publication["_id"])
             pbList.append(publication)
@@ -55,17 +61,21 @@ async def getAllPb():
         raise exception
 
 @router.get("/term/{term}")
-async def getTermPb(term:str):
+async def getTermPb(term:str, auth:User = Depends(verifyTokenAccess)):
     try:
-        pbList = List[Publication]
-        pbList = []
+        pbList:List[Publication] = []
         publications = db.publications.find(
             {
-                "$or":[
-                    {"title":{"$regex":term, "$options":"i"}},
-                    {"content":{"$regex":term, "$options":"i"}},
-                    {"category":{"$regex":term, "$options":"i"}}
+            "$and": [
+                {"userId": auth.id},
+                {
+                "$or": [
+                    {"title": {"$regex": term, "$options": "i"}},
+                    {"content": {"$regex": term, "$options": "i"}},
+                    {"category": {"$regex": term, "$options": "i"}}
                 ]
+                }
+            ]
             }
         )
         for publication in publications:
@@ -76,24 +86,26 @@ async def getTermPb(term:str):
         raise exception
 
 @router.put("/update/{id}", status_code=status.HTTP_200_OK)
-async def updatePb(id:str, title:str  = Form(None), content:str  = Form(None), category:str  = Form(None), tags:Optional[List[str]]  = Form(None)):
+async def updatePb(id:str, title:str  = Form(None), content:str  = Form(None), category:str  = Form(None), tags:Optional[List[str]]  = Form(None), auth:User = Depends(verifyTokenAccess)):
     try:
         publicationDb = srcPublicationId(ObjectId(id))
-        updatedPb = db.publications.find_one_and_replace(
-            {"_id":ObjectId(id)},
-            {
-                "title": title if title is not None else publicationDb["title"],
-                "content": content if content is not None else publicationDb["content"],
-                "category": category if category is not None else publicationDb["category"],
-                "tags": tags if tags is not None else publicationDb["tags"],
-                "createdAt":publicationDb["createdAt"],
-                "updatedAt": datetime.now(timezone.utc).isoformat(timespec='seconds')
-            },
-            return_document=True
-        )
-        
-        return Publication(**publicationSchema(updatedPb))
-    
+        if publicationDb["userId"] == auth.id:
+            updatedPb = db.publications.find_one_and_replace(
+                {"_id":ObjectId(id)},
+                {
+                    "title": title if title is not None else publicationDb["title"],
+                    "content": content if content is not None else publicationDb["content"],
+                    "category": category if category is not None else publicationDb["category"],
+                    "tags": tags if tags is not None else publicationDb["tags"],
+                    "createdAt":publicationDb["createdAt"],
+                    "updatedAt": datetime.now(timezone.utc).isoformat(timespec='seconds')
+                },
+                return_document=True
+            )
+            
+            return Publication(**publicationSchema(updatedPb))
+        else:
+            raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to update this publication")
     except HTTPException as exception:
         if exception.status_code == status.HTTP_404_NOT_FOUND:
             raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Publication not found")
@@ -101,9 +113,13 @@ async def updatePb(id:str, title:str  = Form(None), content:str  = Form(None), c
             raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail={"message":"error: cannot update de publication"})
 
 @router.delete("/delete/{id}", status_code=status.HTTP_204_NO_CONTENT)
-async def deletePb(id:str):
+async def deletePb(id:str, auth:User = Depends(verifyTokenAccess)):
     try:
-        db.publications.delete_one({"_id":ObjectId(id)})
+        publicationDb = db.publications.find_one({"_id":ObjectId(id)})
+        if publicationDb["userId"] == auth.id:
+            db.publications.delete_one({"_id":ObjectId(id)})
+        else:
+            raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to delete this publication")
     except HTTPException as exception:
         if exception.status_code == status.HTTP_404_NOT_FOUND:
             raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Publication not found")
